Producing JSON data via REST proxy
Content-Type is in the format application/vnd.kafka[.embedded_format].[api_version]+[serialization_format]
embedded_format is how the data destined for Kafka is formatted. Must be one of binary, json, or avro
api_version is the API version for REST Proxy -- this should always be v2 as of this writing
serialization_format has nothing to do with your Kafka data, this is how the actual data being sent to REST proxy is serialized. Only json is supported for now -- so always set this to json!